testing list command
